Category items & labels (#27)
* Making item cats and labels standardised.

The html components now rely on the yml data to populate the details such as the text, style and links. Some new parameters allow these can be overridden for use in overlay views.

* Update categoryLabels.yml

* Update categoryItem.html

* Update categoryItem.html

* Update icons.svg

adding downstream infra icon

* Mapped the descriptive text for each category to data file.

Renamed the categoryItems.yml to carbonStandard.yml to encapsulate all the things in the standard.

* Removed CatD Infrastracture

Removed for now until we have some content.